Fixed parsing of same image with new structure

+++ iOSGen/iosgenerator.py
@@ -9,6 +9,7 @@ from colorama import Fore, Back, Style
 from collections import OrderedDict
 import json
 import configparser
+from shutil import copy
 
 """
 listenM = {
@@ -18,20 +19,20 @@ listenM = {
 }
 
 {
-'#0228d4': [['dambakk'], ['self.NAME = [[UIDambakk alloc] initWithFrame:CGRectMake(posX, posY, width, height)];'], ['self.NAME.backgroundColor = [UIColor COLOR];']], 
-'#9aaa01': [['button'], ['self.NAME = [[UIButton alloc] initWithFrame:CGRectMake(posX, posY, width, height)];'], ['self.NAME.backgroundColor = [UIColor COLOR];']], 
+'#0228d4': [['dambakk'], ['self.NAME = [[UIDambakk alloc] initWithFrame:CGRectMake(posX, posY, width, height)];'], ['self.NAME.backgroundColor = [UIColor COLOR];']],
+'#9aaa01': [['button'], ['self.NAME = [[UIButton alloc] initWithFrame:CGRectMake(posX, posY, width, height)];'], ['self.NAME.backgroundColor = [UIColor COLOR];']],
 '#d41d01': [['view'], ['self.NAME = [[UIView alloc] initWithFrame:CGRectMake(posX, posY, width, height)];'], ['self.NAME.backgroundColor = [UIColor COLOR];']]
-} 
+}
 
 listenMP = {
-"default" : ["[self.view addSubview:self.NAME];"],  
-'#d41d01' : ["[self.viewID addSubview:self.NAME];"], 
+"default" : ["[self.view addSubview:self.NAME];"],
+'#d41d01' : ["[self.viewID addSubview:self.NAME];"],
 '#9aaa01' : ["[self.buttonID addSubview:self.NAME];"],
 '#0228d4' : ["[self.dambakkID addSubview:self.NAME;"]
 }
 
 listenH = {
-'#d41d01' : [["view"], ["@property (nonatomic, strong) UIView *NAME;"]], 
+'#d41d01' : [["view"], ["@property (nonatomic, strong) UIView *NAME;"]],
 '#9aaa01' : [["button"], ["@property (nonatomic, strong) UIButton *NAME;"]],
 '#0228d4' : [["Dambakk"], ["@property (nonatomic, superstrong) UIDambakk *NAME;"]]
 }
@@ -42,17 +43,61 @@ listenMP = {}
 listenH = {}
 
 def copyProject(outputPath, args):
-    if(args.ios):
+    if(args.ios and os.path.exists(outputPath)):
+
+        # Finner ID'ene i ny og gammel JSON
+        newJSON, newJSONList = findIDJSON(args.jsonPath)
+        print(newJSON, " test")
+
+        oldJSONPath = outputPath+ "/exampleStructure1.json"
+        oldJSON, oldJSONList = findIDJSON(oldJSONPath)
+        print(oldJSON, " test1")
+
+        # Lager en liste med ID'er som er annerledes, blir flere hvis det er flere her, blir tom hvis ikke.
+        finalJSONID = []
+        for i in newJSON:
+            if(i not in oldJSON):
+                 finalJSONID.append(i)
+
+        print(finalJSONID)
+
+        #print(newJSONList, " new list")
+        print("    ")
+        print("     ")
+        #print(oldJSONList, " old JSON list")
+        print(" ")
+        print(" ")
+
+        # Henter ut en flat struktur, så hvert JSON object kan hentes ut fra en liste, så man slipper reqursion for å hente ut.
+        newFlatJSONList = []
+        for x in newJSONList.items():
+            readElementForNewListe(x[1][1], newFlatJSONList)
+
+        # Gjør samme med gammel JSON struktur
+        oldFlatJSONList = []
+        for y in oldJSONList.items():
+            readElementForNewListe(y[1][1], oldFlatJSONList)
+
+        #Funksjon for å sjekke om vi har riktige ID'er
+        checkIfCorrectID(newFlatJSONList, oldFlatJSONList, finalJSONID)
+
+        print(newFlatJSONList)
+
+    else:
+
+        copy(args.jsonPath, "DemoApp")
         fromDirectory = "DemoApp"
         toDirectory = outputPath
         print("Output path: ", outputPath)
         copy_tree(fromDirectory, toDirectory)
+
 
         configSetup()
 
         newListeM = []
         newListeH = []
         newListeMP = []
+
 
         with open(args.jsonPath) as data_file:
             data = OrderedDict()
@@ -70,7 +115,72 @@ def copyProject(outputPath, args):
         replaceText(outputPath, "m", newListeM, newListeMP)
         replaceText(outputPath, "h", newListeH, None)
 
-        print("Done Dambikk") # haha
+        print("Done") # haha
+
+
+
+def checkIfCorrectID(newJSON, oldJSON, finalJSONID):
+
+    for i in newJSON:
+        tempListe = []
+        for j in oldJSON:
+
+            if(i[1] == j[1] and i[2] == j[2] and i[3] == j[3] and i[4] == j[4] and i[5] == j[5]):
+                i[0] = j[0]
+                oldJSON.remove(j)
+                break
+            elif(i[1] == j[1]):
+                tempListe.append(j)
+
+        if tempListe:
+            match = percentageMatch(tempListe, i)
+            i[0] = match[0]
+            oldJSON.remove(match)
+        elif finalJSONID:
+            print("We came here")
+            i[0] = finalJSONID.pop(0)
+
+
+        print(tempListe)
+
+def percentageMatch(tempListe, elementToCheck):
+    NUMBER = 200
+    TRESHOLD = 0.1
+    for currentElement in tempListe:
+        percentage = abs((currentElement[2]-elementToCheck[2])/NUMBER) + abs((currentElement[3]-elementToCheck[3])/NUMBER) + abs((currentElement[4]-elementToCheck[4])/NUMBER) + abs((currentElement[5]-elementToCheck[5])/NUMBER)
+        currentElement.append(percentage)
+        print(currentElement)
+
+    tempListe.sort(key=lambda x: x[8])
+
+    print(tempListe)
+
+    if(tempListe[0][8] < TRESHOLD):
+        del tempListe[0][-1]
+        print(tempListe[0])
+        return tempListe[0]
+    else:
+        return None
+
+
+def findIDJSON(filepath):
+    test = []
+    newData = []
+    with open(filepath) as data_file:
+        data = OrderedDict()
+        try:
+            data = json.load(data_file, object_pairs_hook=OrderedDict)
+        except ValueError:
+        	print(Fore.RED + "Something went wrong. Could not read JSON. Is the JSON structure correct?" + Style.RESET_ALL)
+        	return
+
+        for e in data.items():
+            readElementForID(e[1][1], test)
+
+
+    return test, data
+
+
 
 def configSetup():
     Config = configparser.RawConfigParser()
@@ -117,7 +227,7 @@ def saveIOSobjectM(color, elementId, posX, posY, width, height, liste, newListeM
             if ("COLOR" in item[0]):
                 (r,g,b) = hex_to_rgb(color)
                 rgbColor = "colorWithRed:{0}/255.0 green:{1}/255.0 blue:{2}/255.0 alpha:1".format(r,g,b)
-                newItem = str(newItem).replace("COLOR", rgbColor) 
+                newItem = str(newItem).replace("COLOR", rgbColor)
             newListeM.append(newItem)
     else:
         print("Not a valid color")
@@ -150,8 +260,34 @@ def saveIOSobjectMP(newListeMP, parentColor, elementId, parent, listenMP, color)
                 newItem = str(item[0]).replace("NAME", str(name))
             newListeMP.append(newItem)
 
+def readElementForNewListe(element, newListe):
+        elementId = element['id']
+        contentStructure = element["content"]
+        color = element["color"]
+        posX = element["x"]
+        posY = element["y"]
+        width = element["width"]
+        height = element["height"]
+        parent = element["parent"]
+        parentColor = element['parentColor']
+        content = "Lorem"
+
+        newListe.append([elementId, color, posX, posY, width, height, parent, parentColor])
+
+        if len(contentStructure) > 0:
+            for i in range (0, len(contentStructure)):
+                content = readElementForNewListe(contentStructure[str(i)][1], newListe)
 
 
+def readElementForID(element, test):
+        elementId = element['id']
+        test.append(elementId)
+        contentStructure = element["content"]
+        content = "Lorem"
+
+        if len(contentStructure) > 0:
+            for i in range (0, len(contentStructure)):
+                content = readElementForID(contentStructure[str(i)][1], test)
 
 def readElement(element, newListeM, newListeH, newListeMP):
     elementId = element['id']
@@ -168,11 +304,12 @@ def readElement(element, newListeM, newListeH, newListeMP):
     saveIOSobjectH(color, elementId,listenH, newListeH)
     saveIOSobjectMP(newListeMP, parentColor, elementId, parent, listenMP, color)
 
+
     content = "Lorem"
     if len(contentStructure) > 0:
         for i in range (0, len(contentStructure)):
             content = readElement(contentStructure[str(i)][1], newListeM, newListeH, newListeMP)
-      
+
 
 def replaceText(templatePath, fileType, elements, elements2):
     templatePath = templatePath+"/DemoApp"
@@ -217,5 +354,5 @@ if __name__== "__main__":
     platform.add_argument("--android", help="Create an android project", action="store_true")
     ap.add_argument("-v", "--verbose", help="Verbose output", action="store_true", default=False)
     args = ap.parse_args()
-    
+
     copyProject(args.outputPath, args)
